refactor(airc-transport): split lan_tcp::adapter into module (one concern per file)
adapter.rs was ~900 lines after the multi-peer changes. Splits into:

  * error.rs (90 LOC) — LanTcpError + Display/Error/From impls
  * inner.rs (63 LOC) — Inner struct + SubscriberHandle + OutboundTx
                       type + per-connection constants
  * connection.rs (166 LOC) — per-connection lifecycle: handle_*,
                              resolve_peer_from_*, install_and_spawn_loops,
                              read_loop, write_loop
  * dispatch.rs (76 LOC) — dispatch_to_subscribers +
                          subscription_matches_with_cursor
  * mod.rs (507 LOC, ~250 tests) — LanTcpAdapter public surface +
                                   Transport impl + integration tests

All-private internals exposed to siblings via `pub(super)`. No public
API changes; tests pass unchanged (35 tests in airc-transport).

Tests: 35 pass, clippy clean (-D warnings), fmt clean.

Targets rust-rewrite directly. One concern per file.
